Re-re:re:fw:re factor some stuff around
