+++ PlayerSync/WebAPI/SignalR/ApiController.Functions.JsonData.cs
@@ -1,7 +1,6 @@
 ﻿using MareSynchronos.API.Data.AdditionalTypes;
 using MareSynchronos.API.Data.Enum;
 using MareSynchronos.API.Dto;
-using MareSynchronos.Interop.Utils;
 using MareSynchronos.PlayerData.Pairs;
 using MareSynchronos.Services.Mediator;
 using Microsoft.AspNetCore.SignalR.Client;
